improve reader

+++ BufferSharp/Reader.cs
@@ -46,7 +46,7 @@ namespace BufferSharp {
         }
         public System.UInt16 GetUInt16 () { CheckBuffers();
 
-            System.UInt32 val = BitConverter.ToUInt16(byteBuffer, index); index += 2;
+            System.UInt16 val = BitConverter.ToUInt16(byteBuffer, index); index += 2;
 
             return val;
         }
